fix: Windows test compilation - C4267 warnings and POSIX functions
Fixed Windows-specific compilation errors in test files:

**Size_t to int conversions (C4267):**
- test_quit_counts_only_active_changes.cpp:156 - Cast all_changes_after.size() to int
- test_quit_after_save_scenario.cpp:97 - Cast all_changes.size() to int

**POSIX function compatibility:**
- test_input_source.cpp - Added Windows compatibility for temporary files
  - Replaced mkstemp (POSIX-only) with tmpnam_s on Windows
  - Added platform-specific includes (<io.h>, <unistd.h>)
  - Defined close as _close on Windows via preprocessor macro
  - Used #ifdef _WIN32 guards for cross-platform temp file creation

All test files now compile cleanly on Windows with MSVC /W4 /WX flags.

+++ tests/test_quit_after_save_scenario.cpp
@@ -94,7 +94,7 @@ TEST_F(QuitAfterSaveScenarioTest, ReproduceUserComplaint) {
     // Check unsaved changes count again (like quit dialog does)
     all_changes = uc.get_all_changes();
     int active_after_save = 0;
-    int total_after_save = all_changes.size();
+    int total_after_save = static_cast<int>(all_changes.size());
 
     for (const auto& change : all_changes) {
         if (change.is_active) {
